a little bit of lovin' to make it through the night
I’ve changed some of the microcopy in placeholders and around the time
boxes.

Created an about page and played with some of the copy we had in the
design file. Still needs work.

revised the copy for when there are no plans in PlanController.js

removed superfulous + symbols from add buttons, have tried to make some
revisions to font sizing because they don’t make sense.